LC-maximum-product-of-two-elements-in-an-array [Runtime: N/A; Memory: N/A]

+++ sorting/maximum-product-of-two-elements-in-an-array.cpp
@@ -2,8 +2,8 @@ class Solution {
 public:
     int maxProduct(vector<int>& nums) {
         
-      int smax = INT_MIN;
-      int max = INT_MIN;
+      long long smax = INT_MIN;
+      long long max = INT_MIN;
       int n = nums.size();
 
       for(int i =0;i<n;i++){
